+++ Yuri_Vislobodsky/Module 5 - Annotations and Generics/javase05/pom.xml
@@ -4,8 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>javase05</groupId>
-    <artifactId>javase05</artifactId>
+    <groupId>javase</groupId>
+    <artifactId>javase</artifactId>
     <version>1.0-SNAPSHOT</version>
