+++ Yuri_Vislobodsky/Module 5 - Annotations and Generics/javase05/src/main/java/javase/t01/annotation/Test.java
@@ -1,4 +1,4 @@
-package javase05.t01.annotation;
+package javase.t01.annotation;
 
 import java.lang.annotation.*;
