+++ Yuri_Vislobodsky/Module 5 - Annotations and Generics/javase05/src/main/java/javase/t01/demo/TesterDemo.java
@@ -1,16 +1,19 @@
-package javase05.t01.demo;
+package javase.t01.demo;
 
-import javase05.t01.testclass.Operation;
-import javase05.t01.testclass.TestClass;
-import javase05.t01.tester.Tester;
+import javase.t01.testclass.Operation;
+import javase.t01.testclass.TestClass;
+import javase.t01.tester.Tester;
 
 /**
  * TesterDemo class runs analysis of TestClass with reflection
  * Created by Yury Vislobodsky on 24.03.2016.
  */
 public class TesterDemo {
+    private TesterDemo() {
+    }
+
     public static void main(String[] args) {
         Tester tester = new Tester();
-        tester.Analyze(new TestClass(2, Operation.DIVIDE, 0));
+        tester.analyze(new TestClass(2, Operation.DIVIDE, 0));
     }
 }
