+++ Yuri_Vislobodsky/Module 5 - Annotations and Generics/javase05/src/main/java/javase/t02/demo/CompareUtilsDemo.java
@@ -1,6 +1,6 @@
-package javase05.t02.demo;
+package javase.t02.demo;
 
-import javase05.t02.compare.CompareUtils;
+import javase.t02.compare.CompareUtils;
 import java.util.Arrays;
 import java.util.Comparator;
