+++ Yuri_Vislobodsky/Module 5 - Annotations and Generics/javase05/src/main/java/javase/t03/demo/TuplesDemo.java
@@ -1,12 +1,15 @@
-package javase05.t03.demo;
+package javase.t03.demo;
 
-import javase05.t03.tuples.*;
+import javase.t03.tuples.*;
 
 /**
  * Demo class for tuples
  * Created by Yury Vislobodsky on 27.03.2016.
  */
 public class TuplesDemo {
+    private TuplesDemo() {
+    }
+
     public static void main(String[] args) {
         Pair<String,Triplet<Integer,Character,Double>> pair =
                 Tuples.create("Hello", Tuples.create(1,'A',3.0));
